requirements.txt
customtkinter-fix-locale-scaling==5.2.2
mutagen==1.47.0
Pillow==12.0.0
pygame==2.6.1
spotdl==4.4.3
